subvalue function estimation program
I need to go back into the math and see what is going on.

Orders 1 and 2 result in a bound that is around 0.46 everywhere. Order 3 raises the bound a bit. But that doesn't make sense with respect to the 'minimum' taken: the unsafe set should always require zero control effort to crash.